renamed isovar.args to isovar.cli for consistency with other py genomics/immunology packages
