+++ script/isovar-allele-counts.py
@@ -20,7 +20,7 @@ Prints number of reads supporting ref, alt, and other alleles at variant loci.
 
 from __future__ import print_function, division, absolute_import
 
-from isovar.args.rna_reads import (
+from isovar.cli.rna_reads import (
     make_rna_reads_arg_parser,
     allele_reads_generator_from_args
 )
